Use JS loader and export contents. (#14)
* Use JS loader and export contents.
Line breaks should be retained when either (a) not minifying, or (b) preserving legal comments. This switches the esbuild loader to `"js"` and creates the returned contents a template literal instead of string. It also updates tests to conform to the template literal implementation.

* Update ci.yml

* Remove unnecessary replace

* Update pnpm-lock.yaml

* Version 1.3.2

* Defer to esbuild initial options.
Users can already specifiy `minify` and/or `legalComments` in their esbuild config. To reduce the amount of setup they have to do, this changes the plugin options to defer to esbuild options, if specified, for minification and legal comment preservation.

* Apply suggestions from code review

Use plain text instead of @defaultValue attributes to describe default behavior of parameters.

Co-authored-by: Raoul v. R. <vanruesc@users.noreply.github.com>

---------

Co-authored-by: Raoul v. R. <vanruesc@outlook.de>
Co-authored-by: Raoul v. R. <vanruesc@users.noreply.github.com>